pyproject.toml
[project]
name = "fastapi_project"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "Waket Zheng", email = "waketzheng@gmail.com" },
]
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "tortoise-orm>=0.25.0",
    "asyncpg>=0.30.0",
    "aerich[toml]>=0.9.0",
    "loguru>=0.7.3",
]
requires-python = "==3.12.*"
readme = "README.md"

[project.license]
text = "MIT"

[tool.pdm]
distribution = false

[tool.aerich]
tortoise_orm = "app.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[dependency-groups]
dev = [
    "fast-dev-cli>=0.15.0",
]
